Move dump command helper func/type
